pyproject.toml
[tool.poetry]
name = "Miku"
version = "0.1.0"
description = "A Continuation of Isaac-To's Miku Discord Bot"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {version = "2.0.0", extras = ["voice"]}
python-dotenv = "*"
aiodns = "*"
Brotli = "*"
SQLAlchemy = "*"
orjson = "*"
lxml = "*"
asyncpraw = "*"
asyncpg = "*"
uvloop = "*"
qrcode = {extras = ["qil"], version = "*"}
cchardet = "*"
beautifulsoup4 = "*"
pysimdjson = "*"
disquest-utils = "*"
async-google-trans-new = "*"
gql = {extras = ["aiohttp"], version = "*"}
rin-exceptions = "*"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
